backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "FastAPI backend — retrieval, reranking, generation"
requires-python = ">=3.12"
dependencies = [
  # Web framework
  "fastapi==0.136.1",
  "uvicorn[standard]==0.34.3",
  "pydantic==2.11.4",
  "pydantic-settings==2.9.1",
  # LLM Orchestration
  "llama-index==0.14.21",
  "llama-index-llms-litellm>=0.4.0",
  "llama-index-embeddings-openai>=0.3.0",
  "litellm>=1.68.0",
  "openai>=1.77.0",
  # Retrieval
  "qdrant-client==1.14.3",
  "elasticsearch==8.17.2",
  # Reranking
  "sentence-transformers==3.4.1",
  "torch>=2.6.0",
  # Database & Cache
  "asyncpg==0.31.0",
  "sqlalchemy[asyncio]==2.0.41",
  "alembic==1.16.1",
  "redis==7.4.0",
  # Observability
  "langfuse==3.9.0",
  "opentelemetry-sdk==1.33.1",
  "opentelemetry-instrumentation-fastapi==0.54b1",
  "opentelemetry-exporter-otlp==1.33.1",
  "prometheus-fastapi-instrumentator==7.1.0",
  # Utils
  "httpx==0.28.1",
  "tenacity==9.1.2",
  "structlog==25.3.0",
  "python-multipart==0.0.20",
  "python-dotenv==1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
